+++ lib/callable_tree/node/internal.rb
@@ -42,7 +42,7 @@ module CallableTree
         if recursive
           child_nodes
             .lazy
-            .select { |node| node.is_a?(Internal) }
+            .select { |node| node.internal? }
             .map { |node| node.find(recursive: true, &block) }
             .reject(&:nil?)
             .first
@@ -58,7 +58,7 @@ module CallableTree
 
         if recursive
           child_nodes.each do |node|
-            node.reject!(recursive: true, &block) if node.is_a?(Internal)
+            node.reject!(recursive: true, &block) if node.internal?
           end
         end
 
@@ -73,7 +73,7 @@ module CallableTree
         reject!(&block) if block_given?
 
         reject! do |node|
-          node.is_a?(Internal) && node.shake!(&block).child_nodes.empty?
+          node.internal? && node.shake!(&block).child_nodes.empty?
         end
       end
 
@@ -161,6 +161,10 @@ module CallableTree
         key = block ? block.call(self) : identity
         value = child_nodes.reduce({}) { |memo, node| memo.merge!(node.outline(&block)) }
         { key => value }
+      end
+
+      def internal?
+        true
       end
 
       protected